+++ resolver.go
@@ -280,6 +280,7 @@ func (r *Resolver) resolveServerAdministrationHandler() *ServerAdministrationHan
 
 	r.serverAdministrationHandler = NewServerAdministrationHandler(
 		r.resolveBaseHandler(),
+		r.ResolveStore(),
 		r.resolveRaceManager(),
 		r.resolveChampionshipManager(),
 		r.resolveServerProcess(),